+++ src/main/java/io/github/torand/openapi2java/writers/java/JavaPojoWriter.java
@@ -16,7 +16,7 @@
 package io.github.torand.openapi2java.writers.java;
 
 import io.github.torand.openapi2java.generators.Options;
-import io.github.torand.openapi2java.model.AnnotationInfo;
+import io.github.torand.openapi2java.model.AnnotatedTypeName;
 import io.github.torand.openapi2java.model.PojoInfo;
 import io.github.torand.openapi2java.model.PropertyInfo;
 import io.github.torand.openapi2java.writers.BaseWriter;
@@ -25,11 +25,8 @@ import io.github.torand.openapi2java.writers.PojoWriter;
 import java.io.Writer;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.stream.Stream;
 
 import static io.github.torand.javacommons.collection.CollectionHelper.nonEmpty;
-import static io.github.torand.javacommons.stream.StreamHelper.streamSafely;
-import static java.util.Objects.nonNull;
 import static java.util.function.Predicate.not;
 import static java.util.stream.Collectors.joining;
 
@@ -67,31 +64,7 @@ public class JavaPojoWriter extends BaseWriter implements PojoWriter {
         pojoInfo.properties().forEach(propInfo -> {
             writeNewLine();
             writePropertyAnnotationLines(propInfo);
-
-            writeIndent(1);
-            if (nonNull(propInfo.type().itemType())) {
-                String itemTypeWithAnnotations = Stream.concat(streamSafely(propInfo.type().itemType().annotations()).map(AnnotationInfo::annotation), Stream.of(propInfo.type().itemType().name()))
-                    .collect(joining(" "));
-
-                if (!opts.pojosAsRecords()) {
-                    write("public ");
-                }
-
-                if (nonNull(propInfo.type().keyType())) {
-                    String keyTypeWithAnnotations = Stream.concat(streamSafely(propInfo.type().keyType().annotations()).map(AnnotationInfo::annotation), Stream.of(propInfo.type().keyType().name()))
-                        .collect(joining(" "));
-
-                    write("%s<%s, %s> %s".formatted(propInfo.type().name(), keyTypeWithAnnotations, itemTypeWithAnnotations, propInfo.name()));
-                } else {
-                    write("%s<%s> %s".formatted(propInfo.type().name(), itemTypeWithAnnotations, propInfo.name()));
-                }
-            } else {
-                if (opts.pojosAsRecords()) {
-                    write("%s %s".formatted(propInfo.type().name(), propInfo.name()));
-                } else {
-                    write("public %s %s".formatted(propInfo.type().name(), propInfo.name()));
-                }
-            }
+            writePropertyTypeAndNameLines(propInfo);
 
             if (opts.pojosAsRecords()) {
                 if (propNo.getAndIncrement() < pojoInfo.properties().size()) {
@@ -171,10 +144,22 @@ public class JavaPojoWriter extends BaseWriter implements PojoWriter {
             writeIndent(1);
             writeLine(a.annotation());
         });
-        propInfo.type().annotations().forEach(a -> {
+    }
+
+    private void writePropertyTypeAndNameLines(PropertyInfo propInfo) {
+        AnnotatedTypeName annotatedTypeName = propInfo.type().getAnnotatedFullName();
+
+        annotatedTypeName.annotations().forEach(a -> {
             writeIndent(1);
-            writeLine(a.annotation());
+            writeLine(a);
         });
+
+        writeIndent(1);
+        if (opts.pojosAsRecords()) {
+            write("%s %s".formatted(annotatedTypeName.typeName(), propInfo.name()));
+        } else {
+            write("public %s %s".formatted(annotatedTypeName.typeName(), propInfo.name()));
+        }
     }
 
     private boolean isInPackage(String qualifiedType, String pojoModelSubpackage) {