+++ src/test/resources/expected-output/kotlin/CompoundApi.kt
@@ -12,6 +12,7 @@ import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotEmpty
 import jakarta.validation.constraints.NotNull
+import jakarta.validation.constraints.Size
 import jakarta.ws.rs.Consumes
 import jakarta.ws.rs.GET
 import jakarta.ws.rs.HeaderParam
@@ -130,6 +131,7 @@ interface CompoundApi {
     @Parameter(`in` = HEADER, name = "X-User-ID", description = "Unique user identifier (SHA1 fingerprint)", required = true, schema = Schema(implementation = String::class))
     @Parameter(`in` = QUERY, name = "offset", description = "Offset of first result in returned page", schema = Schema(implementation = Int::class, defaultValue = "0"))
     @Parameter(`in` = QUERY, name = "limit", description = "Number of results in returned page", schema = Schema(implementation = Int::class, defaultValue = "10"))
+    @Parameter(`in` = QUERY, name = "keywords", description = "Keywords to search for", schema = Schema(type = ARRAY, implementation = String::class))
     @APIResponse(responseCode = "200", description = "OK", content = [ Content(mediaType = APPLICATION_JSON, schema = Schema(type = ARRAY, implementation = ProductV1Dto::class)), Content(mediaType = "application/vnd.test.api.product-v1+json", schema = Schema(type = ARRAY, implementation = ProductV1Dto::class)) ])
     @APIResponse(responseCode = "400", description = "Invalid input parameters supplied", content = [ Content(mediaType = APPLICATION_JSON, schema = Schema(implementation = ErrorDto::class)) ])
     @APIResponse(responseCode = "401", description = "Authentication credentials are invalid or missing", content = [ Content(mediaType = APPLICATION_JSON, schema = Schema(implementation = ErrorDto::class)) ])
@@ -140,7 +142,8 @@ interface CompoundApi {
         @HeaderParam(ACCEPT_LANGUAGE) acceptLanguage: String? = null,
         @HeaderParam("X-User-ID") @NotBlank xUserID: String,
         @QueryParam("offset") @Min(0) offset: Int? = null,
-        @QueryParam("limit") @Min(1) limit: Int? = null
+        @QueryParam("limit") @Min(1) limit: Int? = null,
+        @QueryParam("keywords") @Valid @Size(min = 1) keywords: List<@NotBlank @Size(min = 3) String>? = null
     ): Response
 
     @POST